requirements.txt
amqp==2.6.1
asgiref==3.5.1
async-timeout==4.0.2
billiard==3.6.4.0
braintree==4.15.2
Brotli==1.0.9
celery==4.4.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cssselect2==0.6.0
Deprecated==1.2.13
Django==3.2.13
django-localflavor==3.1
django-parler==2.3
django-rosetta==0.9.8
fonttools==4.33.3
html5lib==1.1
idna==3.3
kombu==4.6.11
packaging==21.3
Pillow==9.1.0
polib==1.1.1
prompt-toolkit==3.0.29
pycparser==2.21
pydyf==0.1.2
pyparsing==3.0.9
pyphen==0.12.0
python-stdnum==1.17
pytz==2022.1
redis==4.3.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tinycss2==1.1.1
urllib3==1.26.9
vine==1.3.0
wcwidth==0.2.5
weasyprint==54.3
webencodings==0.5.1
wrapt==1.14.1
zopfli==0.2.1
